refactor: Move sample configs to src/sampleConfig
